Alter polling processes for machinesDB #5 #7 #12 #16
Polling process now updates machines using a controller rather than through the polling service allowing for the timer to continue countdown without dropping the database. Opens up ability to alter machines when the countdown finishes, but unable to initialize that without ability to modify machines so that they can be changed from running to not running.