change to Client side user interface.
adds color palette for UMN colors, alters the app.component.html and .scss file so the theme is associated with the palette. rearranges the GAuth and HLink for the reporting of machines to be at the top and swapped by position. finally, Name of Website has been changed to Morris Laundry Page.